requirements.txt
pandas
openpyxl
matplotlib
reportlab
google-generativeai
pytest